# packages/contracts/openapi/organization.yaml
openapi: 3.1.0
info:
  title: 联软CRM V3组织架构与企微目录同步接口契约
  version: 2026-08-26
  description: |
    本契约描述当前已挂载的组织架构、角色授权和企业微信目录同步接口。
    所有接口使用同源签名 Cookie 会话，且当前实现只允许 superadmin 访问；
    不接受 operatorId、用户名、兼容令牌或其他客户端自报身份作为授权依据。
    组织功能由 V3_ORGANIZATION_ENABLED 控制，写入还受 V3_ORGANIZATION_WRITE_ENABLED 控制；
    账号入口整合另由 V3_ORGANIZATION_ACCOUNT_ENTRY_MERGED 控制，且仅影响管理端导航呈现；
    停用账号旧会话防护受 V3_AUTH_ACCOUNT_STATUS_CHECK_ENABLED 独立控制；
    停用归档交接额外受 V3_ORGANIZATION_OFFBOARDING_ENABLED 控制，且开启前必须先开启账号状态防护；
    企微目录同步额外受 V3_DIRECTORY_SYNC_ENABLED 控制。以上开关默认均为 false。
    企微目录同步首期仅创建和读取 integration 域预览数据，绝不改写组织、账号、渠道或 CRM 主体数据。
    企业微信身份映射导入仅写入 iam.external_identities，必须由超级管理员上传、预览并确认；不改变企微目录同步、泛微 OA、订单或自动拉群规则。
servers:
  - url: /
    description: V3 同源服务入口
security:
  - V3签名会话: []
...
paths:
  /api/org/status:
    get:
      tags: [组织架构]
      operationId: 读取组织功能状态
      summary: 读取组织架构功能开关状态
      description: 功能关闭时返回 enabled=false；功能开启后要求 superadmin 会话。
      responses:
        "200":
          description: 当前开关状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/组织状态成功响应"
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织无权限" }
...
  /api/org/units/tree:
    get:
      tags: [组织架构]
      operationId: 查询组织树
      summary: 查询组织树
      x-运行条件: V3_ORGANIZATION_ENABLED=true
      responses:
        "200": { $ref: "#/components/responses/组织树成功" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织无权限" }
        "503": { $ref: "#/components/responses/组织功能关闭" }
  /api/org/users/{userId}/eteams-identity:
    get:
      tags: [泛微 OA 身份]
      operationId: 查询泛微OA身份
      summary: 查询账号的泛微 OA 正式身份与全部核验候选
      description: 仅超级管理员可调用。正式身份来自 iam.external_identities；候选来自独立表且不得用于发起 OA 流程。
      parameters: [{ $ref: "#/components/parameters/用户编号" }]
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织无权限" }
        "404": { $ref: "#/components/responses/组织不存在" }
        "503": { $ref: "#/components/responses/组织功能关闭" }
  /api/org/users/{userId}/eteams-identity-candidates:
    post:
      tags: [泛微 OA 身份]
      operationId: 创建泛微OA身份候选
      summary: 录入待人工核验的泛微 OA 身份候选
      description: 写入需同时开启组织功能和组织写入开关。不得按姓名自动创建候选或正式映射；本接口只接受人工明确录入的候选。
      parameters:
        - $ref: "#/components/parameters/用户编号"
        - $ref: "#/components/parameters/幂等键"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/泛微OA身份候选请求" }
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/组织并发或约束冲突" }
        "503": { $ref: "#/components/responses/组织功能关闭" }
  /api/org/users/{userId}/eteams-identity-candidates/{candidateId}:
    put:
      tags: [泛微 OA 身份]
      operationId: 更正泛微OA身份候选
      summary: 更正仍处于待核验状态的候选
      description: 已确认、已驳回或已替代的候选只读保留，禁止覆盖。
      parameters:
        - $ref: "#/components/parameters/用户编号"
        - $ref: "#/components/parameters/候选编号"
        - $ref: "#/components/parameters/幂等键"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/泛微OA身份候选更新请求" }
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/组织并发或约束冲突" }
  /api/org/users/{userId}/eteams-identity-candidates/{candidateId}/confirm:
    post:
      tags: [泛微 OA 身份]
      operationId: 确认泛微OA身份候选
      summary: 原子确认候选并创建泛微 OA 正式身份
      description: 服务端同时校验用户与外部 userid 的有效映射唯一性；冲突时拒绝覆盖。确认成功后，该身份才可作为后续 OA 发起人。
      parameters:
        - $ref: "#/components/parameters/用户编号"
        - $ref: "#/components/parameters/候选编号"
        - $ref: "#/components/parameters/幂等键"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/泛微OA身份确认请求" }
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/组织并发或约束冲突" }
  /api/org/users/{userId}/eteams-identity-candidates/{candidateId}/reject:
    post:
      tags: [泛微 OA 身份]
      operationId: 驳回泛微OA身份候选
      summary: 驳回错误或无法确认的候选
      parameters:
        - $ref: "#/components/parameters/用户编号"
        - $ref: "#/components/parameters/候选编号"
        - $ref: "#/components/parameters/幂等键"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/泛微OA身份驳回请求" }
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/组织并发或约束冲突" }
  /api/org/users/{userId}/eteams-identity/disable:
    post:
      tags: [泛微 OA 身份]
      operationId: 停用泛微OA正式身份
      summary: 停用错误或过期的泛微 OA 正式映射
      description: 不物理删除映射；停用后必须重新录入并人工确认候选，才能恢复 OA 发起资格。
      parameters:
        - $ref: "#/components/parameters/用户编号"
        - $ref: "#/components/parameters/幂等键"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/泛微OA正式身份停用请求" }
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "404": { $ref: "#/components/responses/组织不存在" }
        "409": { $ref: "#/components/responses/组织并发或约束冲突" }
  /api/org/wecom-identities/import-preview:
    post:
      tags: [企业微信身份映射]
      operationId: 预览企业微信身份映射导入
      summary: 上传并预览企业微信 UserId 映射文件
      description: 仅超级管理员且组织写入开关开启时可调用。支持 UTF-8 CSV 和 XLSX；服务端按姓名严格匹配启用 V3 用户，不写入数据库。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/企业微信身份映射文件上传" }
      responses:
        "200": { $ref: "#/components/responses/企业微信身份映射预览成功" }
        "400": { $ref: "#/components/responses/企业微信身份映射请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/组织并发或约束冲突" }
  /api/org/wecom-identities/import-confirm:
    post:
      tags: [企业微信身份映射]
      operationId: 确认企业微信身份映射导入
      summary: 在串行化事务内确认企业微信 UserId 映射导入
      description: 必须重新上传文件并携带 Idempotency-Key。服务端重新校验数据库现状；存在任何阻断项时整批拒绝，成功后逐条写入审计。
      parameters: [{ $ref: "#/components/parameters/幂等键必填" }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema: { $ref: "#/components/schemas/企业微信身份映射文件上传" }
      responses:
        "200": { $ref: "#/components/responses/企业微信身份映射导入成功" }
        "400": { $ref: "#/components/responses/企业微信身份映射请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/企业微信身份映射冲突" }
  /api/org/account-conflicts:
    get:
      tags: [企业微信身份映射]
      operationId: 查询账号与企业微信映射冲突
      summary: 查询同名不同登录账号清单
      description: 仅超级管理员可调用。只返回启用账号中的同名不同登录账号摘要，供人工选择来源账号和保留账号；不自动归并。
      responses:
        "200": { $ref: "#/components/responses/账号冲突清单成功" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织无权限" }
        "503": { $ref: "#/components/responses/组织功能关闭" }
  /api/org/wecom-identities/correct:
    post:
      tags: [企业微信身份映射]
      operationId: 校正企业微信身份映射
      summary: 停用冲突映射并建立目标账号的有效企业微信身份
      description: 仅超级管理员可调用。必须携带预览中的映射编号和行版本、原因、目标账号二次确认及幂等键；不删除历史映射，不修改 OA、订单或群记录。
      parameters: [{ $ref: "#/components/parameters/幂等键必填" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/企业微信身份校正请求" }
      responses:
        "200": { $ref: "#/components/responses/企业微信身份校正成功" }
        "400": { $ref: "#/components/responses/企业微信身份映射请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/企业微信身份映射冲突" }
  /api/org/channel-members/{id}/profile:
    get:
      tags: [组织架构]
      operationId: 查询统一渠道成员档案
      summary: 查询组织、渠道商员工和企业管理员入口共用的成员档案
      description: 仅接受签名 Cookie 超级管理员会话。企业管理员身份独立返回；业务角色固定为销售或技术；证书按用户返回。
      parameters: [{ $ref: "#/components/parameters/组织编号" }]
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织无权限" }
        "404": { $ref: "#/components/responses/组织不存在" }
        "503": { $ref: "#/components/responses/组织功能关闭" }
    put:
      tags: [组织架构]
      operationId: 保存统一渠道成员档案
      summary: 原子保存成员基本资料、销售技术角色及显式证书授权变化
      description: 不允许修改用户名、账号状态、密码、外部身份、企业管理员身份或系统权限角色。业务角色和证书互不自动联动。
      parameters:
        - $ref: "#/components/parameters/组织编号"
        - $ref: "#/components/parameters/幂等键"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/保存统一渠道成员档案请求" }
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/组织并发或约束冲突" }
        "503": { $ref: "#/components/responses/组织功能关闭" }
  /api/org/units/{id}:
    get:
      tags: [组织架构]
      operationId: 查询组织详情
      summary: 查询单个组织
      parameters: [{ $ref: "#/components/parameters/组织编号" }]
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织无权限" }
        "404": { $ref: "#/components/responses/组织不存在" }
        "503": { $ref: "#/components/responses/组织功能关闭" }
    put:
      tags: [组织架构]
      operationId: 更新组织
      summary: 更新组织名称、父组织和排序
      description: 写入要求组织和写入开关开启。更新必须提供 rowVersion；建议发送 Idempotency-Key，当前服务端兼容未发送该头的既有调用。
      parameters:
        - $ref: "#/components/parameters/组织编号"
        - $ref: "#/components/parameters/幂等键"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/更新组织请求" }
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/组织并发或约束冲突" }
        "503": { $ref: "#/components/responses/组织功能关闭" }
  /api/org/units/{id}/status:
    put:
      tags: [组织架构]
      operationId: 更新组织状态
      summary: 启用、停用或归档组织
      description: 归档守卫由服务端执行；不得由前端推断可执行状态。
      parameters:
        - $ref: "#/components/parameters/组织编号"
        - $ref: "#/components/parameters/幂等键"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/状态更新请求" }
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/组织并发或约束冲突" }
        "503": { $ref: "#/components/responses/组织功能关闭" }
  /api/org/units:
    post:
      tags: [组织架构]
      operationId: 创建组织
      summary: 创建草稿组织
      parameters: [{ $ref: "#/components/parameters/幂等键" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/创建组织请求" }
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/组织并发或约束冲突" }
        "503": { $ref: "#/components/responses/组织功能关闭" }
  /api/org/positions:
    get:
      tags: [组织架构]
      operationId: 查询岗位
      summary: 按可选组织查询岗位
      parameters:
        - name: orgUnitId
          in: query
          schema: { $ref: "#/components/schemas/UUID" }
      responses:
        "200": { $ref: "#/components/responses/列表成功" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织无权限" }
        "503": { $ref: "#/components/responses/组织功能关闭" }
    post:
      tags: [组织架构]
      operationId: 创建岗位
      summary: 创建岗位字典项
      parameters: [{ $ref: "#/components/parameters/幂等键" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/创建岗位请求" }
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/组织并发或约束冲突" }
        "503": { $ref: "#/components/responses/组织功能关闭" }
  /api/org/positions/{id}:
    put:
      tags: [组织架构]
      operationId: 更新岗位
      summary: 更新岗位名称和类别
      parameters:
        - $ref: "#/components/parameters/组织编号"
        - $ref: "#/components/parameters/幂等键"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/更新岗位请求" }
      responses:
        "200": { $ref: "#/components/responses/通用成功" }
        "400": { $ref: "#/components/responses/组织请求不合法" }
        "401": { $ref: "#/components/responses/组织未登录" }
        "403": { $ref: "#/components/responses/组织写入关闭" }
        "409": { $ref: "#/components/responses/组织并发或约束冲突" }
...
components:
  securitySchemes:
    V3签名会话:
      type: apiKey
      in: cookie
      name: v3_session
      description: V3 HttpOnly 签名会话 Cookie；具体 Cookie 名称由部署配置维护，客户端不得伪造。
...
  schemas:
...
    组织状态成功响应:
      allOf:
        - $ref: "#/components/schemas/通用成功响应"
        - type: object
          properties:
            data:
              type: object
              required: [enabled, writeEnabled, channelPhoneEditEnabled, directorySyncEnabled, accountEntryMerged, accountStatusCheckEnabled, offboardingEnabled]
              properties:
                enabled: { type: boolean }
                writeEnabled: { type: boolean }
                channelPhoneEditEnabled:
                  type: boolean
                  description: 是否允许修改渠道成员手机号；IAM/UniSDP 手机号匹配专项回归前必须关闭。
                directorySyncEnabled: { type: boolean }
                accountEntryMerged:
                  type: boolean
                  description: 是否启用“组织与账号”导航；不改变登录、会话或接口授权。
                accountStatusCheckEnabled:
                  type: boolean
                  description: 是否持续校验账号状态并拒绝停用账号的旧会话；执行交接前必须开启，回退交接时不得关闭。
                offboardingEnabled:
                  type: boolean
                  description: 是否允许发起、重试和关闭停用归档交接；关闭时仍允许只读影响预览。
...